microservices-refactor/services/book/requirements.txt
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
injector==0.22.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
pillow==11.2.1
pycparser==2.22
PyMySQL==1.1.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3